requirements.txt
faiss-cpu==1.7.4
google-generativeai==0.8.4
sentence-transformers==3.4.1
nltk==3.9.1
gradio==5.20.1
numpy==1.26.4